some improvements to routing status
also added a useful function to misc.h which returns the singular or plural value of a string.  The function takes an integer count and two strings, if the count is 1, then the first string is returned, otherwise the second string is returned.  This is very handy for for constructing clean sentences in text feedback.

git-svn-id: http://fritzing.googlecode.com/svn/trunk@1555 85f94c43-f63a-0410-9d22-21439860c375