fixing issue #385
Now the resistorSetGenerator script can take an extra -b or --bin= argument.
If you point this to the allParts.fz bin file or any other bin which has the current set of resistors, the generator code will try to match the new resistors to the moduleID's it finds in that bin.

also fixed the other issues concerning resistor generation (<label> and missing layers in the footprint svg)

git-svn-id: http://fritzing.googlecode.com/svn/trunk@2076 85f94c43-f63a-0410-9d22-21439860c375