New DilPackageGenerator!
Works as a command line tool. Check out 'python generator.py --help' to see what it does.
(creates all svg's for breadboard, icon & schematic, picks a svg for pcb, and creates the .fzp file)

The only thing the user (or you) has to do after is you created the part is open it in the parts editor and add properties, tags and name the pins correctly.

git-svn-id: http://fritzing.googlecode.com/svn/trunk@2106 85f94c43-f63a-0410-9d22-21439860c375