Added code to check for updates.  It happens automatically when fritzing starts up; you can also invoke it manually by selecting the "check for update" menu item under the help menu.  If there is either a new main version or interim version available, the versions are displayed on a modeless dialog box, and a link is displayed which will pop up a browser pointing to the appropriate download url.  For automatic checking, the dialog box only appears if new versions are found (and the user hasn't previously been informed about those versions).  If there are new versions available and they are displayed, these are stored in Settings.  So once a user has been told about a new version, the automatic update doesn't keep reminding him every time he starts up Fritzing.  These versions will still appear if the user asks for a manual check.
Right now the dialog only shows the most recent main version (if any) and the most recent interim version (if any).  The dialog is not very well formatted at present.  You can change it by fiddling with updatedialog.css in the resources/styles folder.

Since I've left the version at 0.1.14b, if you want to test the dialog, go to version.cpp and make a change, like setting Version::minorVersion to zero, so that it will look to the app like there are newer versions up on the web.



git-svn-id: http://fritzing.googlecode.com/svn/trunk@2256 85f94c43-f63a-0410-9d22-21439860c375