Unified getSaveFileName and getOpenFileName by putting them into FApplication (which then calls the QFileDialog versions).  If the result from either of these calls is non-null, this is cached (minus the filename part) as the default folder.  Save, Save As, Open, and Export from the main window now all use this mechanism.  I left the parts editor alone for now (though I notice if you click successive images, it goes back to the documents folder, even if you have selected an image from another folder: it looks like m_defaultFolder is never being updated).
I found one repeated bug that was breaking the previous attempt at this:  m_filename was treated as a filename, but most of the time it's actually a full path.  The other big difference from the former scheme is that most save/load/export functions were not updating the default folder.

git-svn-id: http://fritzing.googlecode.com/svn/trunk@2327 85f94c43-f63a-0410-9d22-21439860c375