This is a start at plugging memory leaks.  It turns out visual studio offers a reasonable (if obscure) toolset for finding them--I've set it up so that most anything that includes misc.h will report leaks (if you're running in Windows, Visual Studio, Debug mode).
Most of the big "memory leaks' in Fritzing are benign--that is, we create a whole bunch of objects at startup, and those remain with the program until the end--so it was never really necessary to directly delete them.  However, in order to use the windows toolset, those benign objects had to be deleted, because the toolset couldn't tell the difference between benign objects hanging around and a real memory leak--and the benign objects generated 39 megs of leakage log reportage that was impossible to wade through.   

So all that to say that there are now destructors and other direct deletion methods added to officially clean up the benign leaks.  And there also some code changes to deal with the real leaks I've found so far after getting the benign leaks out of the way.  Swap was probably the leakiest.  I have not yet dealt with the parts editor, and there are leaks in daos.h that I will need Mariano's help with.

git-svn-id: http://fritzing.googlecode.com/svn/trunk@2398 85f94c43-f63a-0410-9d22-21439860c375