Swapping from obsolete to new parts is now working.  All the parts living in the obsolete folder have had the prefix "old " added to their family property.
The implementation is that there is a "select all obsolete parts" item on the parts menu.  Choosing that will find each obsolete part in the sketch and select it.  Then you can fiddle with the selection, and choose another item from the parts menu ,"swap obsolete parts", which will swap all the obsolete parts you've selected. 

The good news is that I've hacked it so that pin counts are preserved.  The bad news is that you may lose such properties as resistance, and pin header gender.

A good way to test this is to open the "all user parts bin" and drag some obsolete parts onto the sketch.  Then try the new way of swapping them.  

git-svn-id: http://fritzing.googlecode.com/svn/trunk@3625 85f94c43-f63a-0410-9d22-21439860c375